fix: move fn get_vp_token and review nits
