+++ Cargo.toml
@@ -9,7 +9,7 @@ tauri-build = "2.0.5"
 did_manager = { git = "https://git@github.com/impierce/did-manager", tag = "v1.0.0-beta.5" }
 jsonwebtoken = "9.3"
 log = "^0.4"
-oid4vc = { git = "https://git@github.com/impierce/openid4vc.git", rev = "b4582de" }
+oid4vc = { git = "https://git@github.com/impierce/openid4vc", rev = "b4582de" }
 rand = "0.8"
 rustls = { version = "0.23", features = ["aws-lc-rs"] }
 rustls-platform-verifier = { version = "0.6" }
