+++ identity-wallet/src/state/credentials/reducers/mod.rs
@@ -3,4 +3,3 @@ pub mod handle_oid4vp_authorization_request;
 pub mod self_issue_credential;
 pub mod send_credential_request;
 pub mod update_credential_metadata;
-pub mod vp_token_payload_prep;
